Move configuration code from `auth/mod.rs` to `auth/config.rs`
Just moving code around.
